Change ruc type from int to Long

+++ src/main/java/com/ProyectoDeAula5/Proyecto5/model/Proveedor.java
@@ -16,7 +16,7 @@ public class Proveedor {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int ruc;
+    private Long ruc;
     private String nombre;
     private String telefono;
     private String direccion;
